requirements.txt
fastapi
uvicorn
python-dotenv
requests
pillow
pdf2image
